refactor(editor): share the other-instance check between editors
The check reads the payload every editor /open route returns, so it does
not belong to OnlyOffice. Excalidraw and the PDF editor need it next.
